+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on a modified version of [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.3.1
+
+### News
+
+* Adds support for callbacks before instance deletion.
+* Improves test coverage.
+
 ## 0.3.0
 
 ### News